utils: dragonboard: restart config pipe service if it fails
This is a trivial service that needs to start successfully in order to have
the aditof-demo running, otherwise a manual run of the config-pipe.sh
script is needed. On some "boots" this service fails to start due to being
unable to open "/dev/media1". In this case the service could try to restart
after a while, and keep restarting until it is successful

Signed-off-by: Daniel Guramulta <daniel.guramulta@analog.com>