python/examples/gesture_rec/notebook.py: Analyze the video continuously
Removed Snap! button and Process tab from the notebook. Changed image
display from label to a separate window, in order to have better frame
rate.

The camera requests only the depth image.

After clicking on Start Video, the depth image is displayed
continuously. Every 1 second a process of finger detection is started on
a different core. After finishing the calculation the bounding box of
the hand and the number of detected fingers is updated.

Signed-off-by: Daniel Matyas <daniel.matyas@analog.com>