python/examples/gesture_rec/process.py: Made process faster.
First of all, it is faster, because there is no GUI, only computation.

Used depth image directly for computation; did not convert image to RGB.

Made max distance of the bit image dynamic. It depends on the closest
pixel.

Changed hand center to the point farthest from the background

Made cdist_threshold dynamic. It is inversely proportional to
the max distance of bit image.

Selected the vertex which is farthest from the hand center on every
finger.

Neglected the vertices which are below the wrist.

Signed-off-by: Daniel Matyas <daniel.matyas@analog.com>